+++ integration-tests/helpers/cart-seeder.js
@@ -904,6 +904,7 @@ module.exports = async (dataSource, data = {}) => {
     unit_price: 8000,
     quantity: 1,
     variant_id: "test-variant",
+    product_id: "test-product",
     cart_id: "test-cart-2",
   })
   await manager.save(li)
@@ -957,6 +958,7 @@ module.exports = async (dataSource, data = {}) => {
     unit_price: 8000,
     quantity: 1,
     variant_id: "test-variant",
+    product_id: "test-product",
     cart_id: "test-cart-3",
   })
   await manager.save(li2)
@@ -1036,6 +1038,7 @@ module.exports = async (dataSource, data = {}) => {
     quantity: 1,
     variant_id: "test-variant-sale-cg",
     cart_id: "test-cart-3",
+    product_id: "test-product",
     metadata: { "some-existing": "prop" },
   })
   await manager.save(li3)