+++ packages/types/src/modules-sdk/index.ts
@@ -124,6 +124,14 @@ export type ModuleJoinerConfig = Omit<
   relationships?: ModuleJoinerRelationship[]
   extends?: {
     serviceName: string
+    fieldAlias?: Record<
+      string,
+      | string
+      | {
+          path: string
+          forwardArgumentsOnPath: string[]
+        }
+    > // alias for deeper nested relationships (e.g. { 'price': 'prices.calculated_price_set.amount' })
     relationship: ModuleJoinerRelationship
   }[]
   serviceName?: string
@@ -168,6 +176,7 @@ export type ModuleJoinerConfig = Omit<
 }
 
 export declare type ModuleJoinerRelationship = JoinerRelationship & {
+  isInternalService?: boolean // If true, the relationship is an internal service from the medusa core TODO: Remove when there are no more "internal" services
   deleteCascade?: boolean // If true, the link joiner will cascade deleting the relationship
 }
 
@@ -201,6 +210,7 @@ export interface ModuleServiceInitializeOptions {
     database?: string
     driverOptions?: Record<string, unknown>
     debug?: boolean
+    pool?: Record<string, unknown>
   }
 }