+++ packages/types/src/pricing/common/price-set.ts
@@ -1,0 +1,32 @@
+import { BaseFilterable } from "../../dal"
+import { FilterableMoneyAmountProps, MoneyAmountDTO } from "./money-amount"
+
+export interface PricingContext {
+  currency_code?: string
+}
+export interface PriceSetDTO {
+  id: string
+  money_amounts?: MoneyAmountDTO[]
+}
+
+export interface CalculatedPriceSetDTO {
+  id: string
+  amount: number | null
+  currency_code: string | null
+  min_quantity: number | null
+  max_quantity: number | null
+}
+
+export interface CreatePriceSetDTO {
+  money_amounts?: MoneyAmountDTO[]
+}
+
+export interface UpdatePriceSetDTO {
+  id: string
+}
+
+export interface FilterablePriceSetProps
+  extends BaseFilterable<FilterablePriceSetProps> {
+  id?: string[]
+  money_amounts?: FilterableMoneyAmountProps
+}