+++ packages/admin/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @medusajs/admin
 
+## 7.0.2
+
+### Patch Changes
+
+- [#4888](https://github.com/medusajs/medusa/pull/4888) [`50798d118`](https://github.com/medusajs/medusa/commit/50798d118986ccd18e1504129773ca40e63abd68) Thanks [@fPolic](https://github.com/fPolic)! - fix(admin-ui): bulk editor cell separators
+
+- [#4950](https://github.com/medusajs/medusa/pull/4950) [`545022145`](https://github.com/medusajs/medusa/commit/545022145b3d882a8f8653f2657b877506e12470) Thanks [@kasperkristensen](https://github.com/kasperkristensen)! - fix(admin): Correctly export RouteProps and SettingProps from @medusajs/admin
+
+- [#4834](https://github.com/medusajs/medusa/pull/4834) [`470379e63`](https://github.com/medusajs/medusa/commit/470379e631ca17ee486a28c50ec828344c61d727) Thanks [@StephixOne](https://github.com/StephixOne)! - fix(medusa, admin-ui): Allow soft-deleted return reason restoration
+
+- Updated dependencies [[`50798d118`](https://github.com/medusajs/medusa/commit/50798d118986ccd18e1504129773ca40e63abd68), [`b7de679e9`](https://github.com/medusajs/medusa/commit/b7de679e9b665368074d406ac5d9b3fbb0187a68), [`470379e63`](https://github.com/medusajs/medusa/commit/470379e631ca17ee486a28c50ec828344c61d727)]:
+  - @medusajs/admin-ui@2.0.2
+
 ## 7.0.1
 
 ### Patch Changes